refactor(global): Adding SQLServer support
- Adding make instructions
- Adding docker-compose for MSSQL
- Fix listing issues, force casting type get auth company_id

+++ app/Http/Controllers/Customer/CustomerIndexController.php
@@ -22,7 +22,7 @@ class CustomerIndexController extends MainController
         if ($request->ajax()) {
             $term = $request->term;
 
-            $customers = Customer::where('company_id', Auth::user()->company_id)
+            $customers = Customer::where('company_id', (int) Auth::user()->company_id)
                 ->where('name', 'LIKE', "%$term%")
                 ->get(['id', 'name as label']);
 
@@ -81,13 +81,13 @@ class CustomerIndexController extends MainController
         $data['colors'] = config('color');
         $data['bootstrap_colors'] = ['text-bg-primary', 'text-bg-secondary', 'text-bg-success', 'text-bg-danger', 'text-bg-warning', 'text-bg-info'];
         $data['countries'] = Country::orderBy('name')->get();
-        $data['customer_count'] = Customer::where('company_id', Auth::user()->company_id)->count();
-        $data['customers'] = $customer->getAllByCompanyId(Auth::user()->company_id, $search, $filters, $order_by);
+        $data['customer_count'] = Customer::where('company_id', (int) Auth::user()->company_id)->count();
+        $data['customers'] = $customer->getAllByCompanyId((int) Auth::user()->company_id, $search, $filters, $order_by);
         $data['search'] = $search;
-        $data['sellers'] = $user->getAllActiveByCompany(Auth::user()->company_id);
+        $data['sellers'] = $user->getAllActiveByCompany((int) Auth::user()->company_id);
         $data['statuses'] = $customer->getStatus();
         // Temporary fix get this from configuration
-        $data['industries'] = (Auth::user()->company_id == 3) ? $industry->getAllByCompany(Auth::user()->company_id) : $industry->getAll();
+        $data['industries'] = ((int) Auth::user()->company_id == 3) ? $industry->getAllByCompany((int) Auth::user()->company_id) : $industry->getAll();
 
         return view('customer.index', $data);
     }
