+++ app/Http/Controllers/Product/ProductIndexController.php
@@ -18,7 +18,7 @@ class ProductIndexController extends MainController
         if ($request->ajax()) {
             $term = $request->term;
 
-            $products = Product::where('company_id', Auth::user()->company_id)
+            $products = Product::where('company_id', (int) Auth::user()->company_id)
                 ->where('name', 'LIKE', "%$term%")
                 ->orWhere('sku', 'LIKE', "%$term%")
                 ->get(['id', 'name as label', 'price']);
@@ -32,10 +32,10 @@ class ProductIndexController extends MainController
         $category = new Category();
         $brand = new Brand();
         $data['search'] = $search;
-        $data['product_count'] = Product::where('company_id', Auth::user()->company_id)->count();
-        $data['products'] = $product->getAllByCompanyId(Auth::user()->company_id, $search, $filters);
-        $data['categories'] = $category->getAllActiveByCompany(Auth::user()->company_id);
-        $data['brands'] = $brand->getAllActiveByCompany(Auth::user()->company_id);
+        $data['product_count'] = Product::where('company_id', (int) Auth::user()->company_id)->count();
+        $data['products'] = $product->getAllByCompanyId((int) Auth::user()->company_id, $search, $filters);
+        $data['categories'] = $category->getAllActiveByCompany((int) Auth::user()->company_id);
+        $data['brands'] = $brand->getAllActiveByCompany((int) Auth::user()->company_id);
 
         return view('product/index', $data);
     }
